Preventing a shortcut which requires permissions from being added to homescreen
A shortcut can be added by any app as INSTALL_SHORTCUT is a normal
level permission. But the intent is actually launched by the launcher
app which can have other permission as well.
> When adding a shortcut from the broadcast, verify that the intent does
not require any permission
> When adding a shortcut using the two-step drop process, verify that
the source app also has the permission to create such a shortcut

Change-Id: I1047823d98597b053c1808df8e51db0b7595ca2a
CRs-Fixed: 1090593